Move phenomenon of renderer out from common by introducing different implementations for environments
