+++ src/renderer/cluster-frame-context/cluster-frame-context.injectable.ts
@@ -5,7 +5,7 @@
 import { getInjectable } from "@ogre-tools/injectable";
 import { ClusterFrameContext } from "./cluster-frame-context";
 import namespaceStoreInjectable from "../components/+namespaces/store.injectable";
-import hostedClusterInjectable from "../../common/cluster-store/hosted-cluster.injectable";
+import hostedClusterInjectable from "./hosted-cluster.injectable";
 import assert from "assert";
 
 const clusterFrameContextInjectable = getInjectable({
